Move drep type parsers to Common module
