+++ cardano-cli/src/Cardano/CLI/EraBased/Options/Governance.hs
@@ -21,7 +21,6 @@ import           Cardano.CLI.Types.Governance
 import           Cardano.CLI.Types.Key
 
 import           Data.Foldable
-import           Data.Functor
 import           Data.Maybe
 import           Data.String
 import           Options.Applicative
@@ -169,49 +168,6 @@ pDRepScriptHash =
 
 scriptHashReader :: ReadM ScriptHash
 scriptHashReader = eitherReader $ Right . fromString
-
-pDRepVerificationKeyOrHashOrFile
-  :: Parser (VerificationKeyOrHashOrFile DRepKey)
-pDRepVerificationKeyOrHashOrFile =
-  asum
-    [ VerificationKeyOrFile <$> pDRepVerificationKeyOrFile
-    , VerificationKeyHash <$> pDRepVerificationKeyHash
-    ]
-
-pDRepVerificationKeyHash :: Parser (Hash DRepKey)
-pDRepVerificationKeyHash =
-    Opt.option (pBech32KeyHash AsDRepKey <|> pHexHash AsDRepKey) $ mconcat
-      [ Opt.long "drep-key-hash"
-      , Opt.metavar "HASH"
-      , Opt.help $ mconcat
-          [ "DRep verification key hash (either Bech32-encoded or hex-encoded).  "
-          , "Zero or more occurences of this option is allowed."
-          ]
-      ]
-
-pDRepVerificationKey :: Parser (VerificationKey DRepKey)
-pDRepVerificationKey =
-  Opt.option (readVerificationKey AsDRepKey) $ mconcat
-    [ Opt.long "drep-verification-key"
-    , Opt.metavar "STRING"
-    , Opt.help "DRep verification key (Bech32 or hex-encoded)."
-    ]
-
-pDRepVerificationKeyOrFile :: Parser (VerificationKeyOrFile DRepKey)
-pDRepVerificationKeyOrFile =
-  asum
-    [ VerificationKeyValue <$> pDRepVerificationKey
-    , VerificationKeyFilePath <$> pDRepVerificationKeyFile
-    ]
-
-pDRepVerificationKeyFile :: Parser (VerificationKeyFile In)
-pDRepVerificationKeyFile =
-  fmap File . Opt.strOption $ mconcat
-    [ Opt.long "drep-verification-key-file"
-    , Opt.metavar "FILE"
-    , Opt.help "Filepath of the DRep verification key."
-    , Opt.completer (Opt.bashCompleter "file")
-    ]
 
 --------------------------------------------------------------------------------
